Bug 572760 Acquire dumps improvements
Fix SpotBugs warnings

Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=572760

+++ plugins/org.eclipse.mat.hprof/src/org/eclipse/mat/hprof/SeekableStream.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019,2020 IBM Corporation and others.
+ * Copyright (c) 2019,2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -688,7 +688,7 @@ public class SeekableStream extends InputStream implements Closeable, AutoClosea
             pos = p.position();
         }
         if (verbose && n > 0 && bestgap < 64)
-            System.out.println("n="+n+" avg="+(gaptotal/n)+" rms="+Math.sqrt(gapstotal/n)+" smallest="+bestgap+" pos="+best.pos); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$
+            System.out.println("n="+n+" avg="+((float)gaptotal/n)+" rms="+Math.sqrt(gapstotal/n)+" smallest="+bestgap+" pos="+best.pos); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$ //$NON-NLS-4$ //$NON-NLS-5$
     }
 
     /**
